requirements.txt
# Essential dependencies only
gpiozero==2.0.1
i2c-lcd==0.1.0
requests==2.32.5
PyYAML>=6.0.0